Delete routes & structuration
