+++ data-provider-API.yaml
@@ -21,7 +21,7 @@ paths:
   /debiai/info:
     get:
       x-eov-operation-id: info
-      x-eov-operation-handler: debiai.js
+      x-eov-operation-handler: info.js
       description: Get some general informations about how DebiAI can use this service
       tags:
         - Info
@@ -54,7 +54,7 @@ paths:
   /debiai/projects:
     get:
       x-eov-operation-id: getProjectsOverview
-      x-eov-operation-handler: debiai.js
+      x-eov-operation-handler: projects.js
       description: Get some general informations about project accessible from this data provider
       tags:
         - Projects
@@ -99,7 +99,7 @@ paths:
         required: true
     get:
       x-eov-operation-id: getProject
-      x-eov-operation-handler: debiai.js
+      x-eov-operation-handler: projects.js
       description: Get Informations about project
       tags:
         - Projects
@@ -178,6 +178,15 @@ paths:
                     type: integer
                     description: Optional number of samples that can be provided by the project for display purpose
                     default: null
+    delete:
+      x-eov-operation-id: deleteProject
+      x-eov-operation-handler: projects.js
+      description: Delete a project
+      tags:
+        - Projects
+      responses:
+        "200":
+          description: Project deleted
   /debiai/projects/{projectId}/data-id-list:
     parameters:
       - in: path
@@ -218,10 +227,9 @@ paths:
         name: analysisEnd
         description: True if it's the last request of an analysis
         required: false
-
     get:
       x-eov-operation-id: dataIdList
-      x-eov-operation-handler: debiai.js
+      x-eov-operation-handler: projects.js
       description: >-
         Get a ID list of all available data in the project, this endpoint need to
         work with the project list provided in the info endoint
@@ -278,7 +286,7 @@ paths:
         required: false
     post:
       x-eov-operation-id: data
-      x-eov-operation-handler: debiai.js
+      x-eov-operation-handler: projects.js
       description: >-
         Get the data to be displayed into DebiAI from the given data ID list,
         the requested IDs will be those given by the endpoint dataIdList. The
@@ -333,7 +341,7 @@ paths:
         required: true
     get:
       x-eov-operation-id: modelList
-      x-eov-operation-handler: debiai.js
+      x-eov-operation-handler: models.js
       description: Tell DebiAI which models are available for the given project
       tags:
         - Models
@@ -379,7 +387,7 @@ paths:
         required: true
     get:
       x-eov-operation-id: modelEvaluatedDataIdList
-      x-eov-operation-handler: debiai.js
+      x-eov-operation-handler: models.js
       description: Tell DebiAI which models are available for the given project
       tags:
         - Models
@@ -415,7 +423,7 @@ paths:
         required: true
     post:
       x-eov-operation-id: modelResults
-      x-eov-operation-handler: debiai.js
+      x-eov-operation-handler: models.js
       description: Gives the results of a model from a given data ID list
       tags:
         - Models
@@ -452,6 +460,33 @@ paths:
                   - type: string
                   - type: number
                   - type: integer
+  /debiai/projects/{projectId}/models/{modelId}:
+    parameters:
+      - schema:
+          type: string
+          minLength: 1
+          example: Project 1
+        name: projectId
+        in: path
+        description: project name to get the data ID list
+        required: true
+      - schema:
+          type: string
+          minLength: 1
+          example: Model 1
+        name: modelId
+        in: path
+        description: Model id to get the data ID list
+        required: true
+    delete:
+      x-eov-operation-id: deleteModel
+      x-eov-operation-handler: models.js
+      description: Delete a model
+      tags:
+        - Models
+      responses:
+        "200":
+          description: The model has been deleted
   ######  ---------------------- SELECTIONS ---------------------- ######
   /debiai/projects/{projectId}/selections:
     parameters:
@@ -465,7 +500,7 @@ paths:
         required: true
     get:
       x-eov-operation-id: selectionList
-      x-eov-operation-handler: debiai.js
+      x-eov-operation-handler: selections.js
       description: Tell DebiAI which selections are available for the given project
       tags:
         - Selections
@@ -493,7 +528,7 @@ paths:
                       default: 0
     post:
       x-eov-operation-id: createSelection
-      x-eov-operation-handler: debiai.js
+      x-eov-operation-handler: selections.js
       description: Create a selection with id List send in body
       tags:
         - Selections
@@ -535,7 +570,7 @@ paths:
         required: true
     get:
       x-eov-operation-id: selectionDataIdList
-      x-eov-operation-handler: debiai.js
+      x-eov-operation-handler: selections.js
       description: Tell DebiAI the samples selected by the a selection
       tags:
         - Selections
@@ -571,7 +606,7 @@ paths:
         required: true
     delete:
       x-eov-operation-id: deleteSelection
-      x-eov-operation-handler: debiai.js
+      x-eov-operation-handler: selections.js
       description: Delete a single selection
       tags:
         - Selections
